+++ stratum/hal/lib/p4/BUILD
@@ -415,9 +415,13 @@ stratum_cc_library(
         ":p4_pipeline_config_cc_proto",
         ":p4_table_map_cc_proto",
         "//stratum/lib:macros",
+        "//stratum/lib:utils",
         "//stratum/public/lib:error",
         "@com_google_absl//absl/strings",
         "@com_google_absl//absl/strings:str_format",
+        "@com_google_googleapis//google/rpc:code_cc_proto",
+        "@com_google_googleapis//google/rpc:status_cc_proto",
+        "@com_github_grpc_grpc//:grpc++",
         "@com_github_p4lang_p4runtime//:p4info_cc_proto",
         "//stratum/glue/gtl:map_util",
         "//stratum/glue/status:statusor",
