+++ stratum/hal/lib/p4/utils.cc
@@ -2,17 +2,19 @@
 // Copyright 2018-present Open Networking Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-
 // Implementation of p4_utils functions.
 
 #include "stratum/hal/lib/p4/utils.h"
 
 #include "absl/strings/str_format.h"
 #include "absl/strings/substitute.h"
+#include "google/rpc/code.pb.h"
+#include "google/rpc/status.pb.h"
 #include "p4/config/v1/p4info.pb.h"
-#include "stratum/lib/macros.h"
-#include "stratum/public/lib/error.h"
 #include "stratum/glue/gtl/map_util.h"
+#include "stratum/lib/macros.h"
+#include "stratum/lib/utils.h"
+#include "stratum/public/lib/error.h"
 
 namespace stratum {
 namespace hal {
@@ -26,8 +28,8 @@ std::string PrintP4ObjectID(int object_id) {
   std::string resource_name =
       ::p4::config::v1::P4Ids::Prefix_Name(resource_type);
   if (resource_name.empty()) resource_name = "INVALID";
-  return absl::StrFormat("%s/0x%x (0x%x)", resource_name.c_str(),
-                         base_id, object_id);
+  return absl::StrFormat("%s/0x%x (0x%x)", resource_name.c_str(), base_id,
+                         object_id);
 }
 
 // This unnamed namespace hides a function that forms a status string to refer
@@ -37,8 +39,8 @@ namespace {
 std::string AddP4ObjectReferenceString(const std::string& log_p4_object) {
   std::string referenced_object;
   if (!log_p4_object.empty()) {
-    referenced_object = absl::Substitute(" referenced by P4 object $0",
-                                         log_p4_object.c_str());
+    referenced_object =
+        absl::Substitute(" referenced by P4 object $0", log_p4_object.c_str());
   }
   return referenced_object;
 }
@@ -55,18 +57,51 @@ std::string AddP4ObjectReferenceString(const std::string& log_p4_object) {
   if (map_value != nullptr) {
     if (map_value->descriptor_case() != descriptor_case) {
       return MAKE_ERROR(ERR_INTERNAL)
-          << "P4PipelineConfig descriptor for " << table_map_key
-          << AddP4ObjectReferenceString(log_p4_object)
-          << " does not have the expected descriptor case: "
-          << map_value->ShortDebugString();
+             << "P4PipelineConfig descriptor for " << table_map_key
+             << AddP4ObjectReferenceString(log_p4_object)
+             << " does not have the expected descriptor case: "
+             << map_value->ShortDebugString();
     }
   } else {
     return MAKE_ERROR(ERR_INTERNAL)
-        << "P4PipelineConfig table map has no descriptor for "
-        << table_map_key << AddP4ObjectReferenceString(log_p4_object);
+           << "P4PipelineConfig table map has no descriptor for "
+           << table_map_key << AddP4ObjectReferenceString(log_p4_object);
   }
 
   return map_value;
+}
+
+std::string P4RuntimeGrpcStatusToString(const ::grpc::Status& status) {
+  std::stringstream ss;
+  if (!status.error_details().empty()) {
+    ss << "(overall error code: "
+       << ::google::rpc::Code_Name(ToGoogleRpcCode(status.error_code()))
+       << ", overall error message: "
+       << (status.error_message().empty() ? "None" : status.error_message())
+       << "). Error details: ";
+    ::google::rpc::Status details;
+    if (!details.ParseFromString(status.error_details())) {
+      ss << "Failed to parse ::google::rpc::Status from GRPC status details.";
+    } else {
+      for (int i = 0; i < details.details_size(); ++i) {
+        ::p4::v1::Error detail;
+        if (details.details(i).UnpackTo(&detail)) {
+          ss << "\n(error #" << i + 1 << ": error code: "
+             << ::google::rpc::Code_Name(ToGoogleRpcCode(detail.code()))
+             << ", error message: "
+             << (detail.message().empty() ? "None" : detail.message()) << ") ";
+        }
+      }
+    }
+  } else {
+    ss << "(error code: "
+       << ::google::rpc::Code_Name(ToGoogleRpcCode(status.error_code()))
+       << ", error message: "
+       << (status.error_message().empty() ? "None" : status.error_message())
+       << ").";
+  }
+
+  return ss.str();
 }
 
 }  // namespace hal
